OWLouvain: Move louvain library to own file, pylint fixes
